refactor: Add 'surname' field to User model and related files
The User model, UserController.php, UserSeeder.php, and related views have been updated to include the 'surname' field. This change allows for storing the surname of users in addition to their name. The necessary modifications have been made to the database migration, validation rules, and form inputs. This refactoring enhances the user management functionality in the application.