refactor: Update main page layout and widgets
This commit updates the main page layout and widgets. The default.blade.php file is added to the views/cms/pages/main/index directory, which extends the cms.master.default layout. The content section is defined, and the magicGrid container is added to wrap the widgets. The widgets for news, categories, images, and users are included in the magicGrid container. This refactor improves the organization and structure of the main page in the CMS.